Add PrepareWebSocketClient and GetProxyOptions options. GetProxyOptions replaces the ProxyOptions in RunProxy.
