+++ src/Microsoft.AspNetCore.Proxy/ProxyApplicationBuilderExtensions.cs
@@ -1,0 +1,25 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using Microsoft.AspNetCore.Proxy;
+
+namespace Microsoft.AspNetCore.Builder
+{
+    public static class ProxyApplicationBuilderExtensions
+    {
+        /// <summary>
+        /// Runs proxy forwarding requests to the server specified by options.
+        /// </summary>
+        /// <param name="app"></param>
+        public static void RunProxy(this IApplicationBuilder app)
+        {
+            if (app == null)
+            {
+                throw new ArgumentNullException(nameof(app));
+            }
+
+            app.UseMiddleware<ProxyMiddleware>();
+        }
+    }
+}
