+++ src/Microsoft.AspNetCore.Proxy/ProxyOptions.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using Microsoft.AspNetCore.Http;
 
 namespace Microsoft.AspNetCore.Proxy
@@ -29,5 +30,27 @@ namespace Microsoft.AspNetCore.Proxy
         /// Query string parameters to append to each request
         /// </summary>
         public QueryString AppendQuery { get; set; }
+
+        /// <summary>
+        /// Creates <see cref="ProxyOptions"/> from a <see cref="Uri"/>
+        /// </summary>
+        /// <param name="baseUri">The <see cref="Uri"/></param>
+        /// <returns>The <see cref="ProxyOptions"/></returns>
+        public static ProxyOptions FromUri(Uri baseUri)
+        {
+            if (baseUri == null)
+            {
+                throw new ArgumentNullException(nameof(baseUri));
+            }
+
+            var options = new ProxyOptions
+            {
+                Scheme = baseUri.Scheme,
+                Host = new HostString(baseUri.Authority),
+                PathBase = baseUri.AbsolutePath,
+                AppendQuery = new QueryString(baseUri.Query)
+            };
+            return options;
+        }
     }
 }
