+++ src/Microsoft.AspNetCore.Proxy/SharedProxyOptions.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Net.Http;
+using System.Net.WebSockets;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 
@@ -21,9 +22,19 @@ namespace Microsoft.AspNetCore.Proxy
         public HttpMessageHandler MessageHandler { get; set; }
 
         /// <summary>
+        /// Function to get the options, where to forward a request to.
+        /// </summary>
+        public Func<HttpRequest, Task<ProxyOptions>> GetProxyOptions { get; set; }
+
+        /// <summary>
         /// Allows to modify HttpRequestMessage before it is sent to the Message Handler.
         /// </summary>
         public Func<HttpRequest, HttpRequestMessage, Task> PrepareRequest { get; set; }
+
+        /// <summary>
+        /// Allows to modify ClientWebSocket before the connection is opened.
+        /// </summary>
+        public Func<HttpRequest, ClientWebSocket, Task> PrepareWebSocketClient { get; set; }
 
         /// <summary>
         /// Keep-alive interval for proxied Web Socket connections.
